backend/photopro/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.31.0
alembic==1.13.3
pydantic==2.8.2
pydantic-settings==2.5.2
httpx==0.27.2
celery==5.4.0
redis==5.1.0
boto3==1.35.0
Pillow==10.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
slowapi==0.1.9
python-multipart==0.0.12
resend==2.3.0
tenacity==9.0.0
zipstream-ng==1.9.0
